chatbot/requirements.txt
Flask
google-cloud-dialogflow
requests
python-dotenv
